Conexion desde Propiedades. Primer consulta Usuario.

+++ entrega/src/Login.cs
@@ -40,7 +40,7 @@ namespace PalcoNet
 
         private void button1_Click(object sender, EventArgs e)
         {
-            int idUsuario = dao.esUsuarioActivo(usuario.Text, pass.Text);
+            Decimal idUsuario = dao.esUsuarioActivo(usuario.Text, pass.Text);
             if (idUsuario <0 )
             {
                 System.Windows.Forms.MessageBox.Show("Credenciales incorrectas para el usuario "+usuario.Text+ ".\nEs posible que el usuario no exista o la contraseña sea incorrecta.\nPor favor, intente de nuevo.");
